fix(data): address TanStack wiring review findings
Review on phase 1 raised one blocker + three follow-ups. All
addressed:

- BLOCKER: QueryClientProvider was outside ErrorBoundary, meaning
  any TanStack-originated render error would bypass Sentry's
  componentDidCatch. Move the provider INTO App.tsx, inside the
  boundary. Devtools move alongside so they read the real query
  cache. main.tsx reverts to the pre-wiring shape.
- IMPORTANT: document in queryClient.ts the post-mutation contract —
  staleTime alone doesn't protect against data written by the
  current session; every useMutation must invalidate its lists via
  queryClient.invalidateQueries in onSuccess / onSettled.
- NEEDS_DISCUSSION: document the latent dependency on the
  AuthContext visibilitychange handler being migrated to call
  queryClient.invalidateQueries() after the 5-min staleness
  threshold. Without that, a user returning to an already-mounted
  page after > 5 min sees unbounded stale data (up to gcTime).
- RECOMMENDED: @tanstack/react-query-devtools moved from
  dependencies to devDependencies. Production installs that use
  --production no longer pull it.

Build + lint + tests still green (315/315). Dev Chrome confirms
the devtools button is still rendered and the app mounts.
