+++ package.json
@@ -29,6 +29,7 @@
     "@biomejs/biome": "^2.4.4",
     "@sentry/vite-plugin": "^5.1.1",
     "@tailwindcss/vite": "^4.1.18",
+    "@tanstack/react-query-devtools": "^5.99.0",
     "@types/node": "^24.10.1",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
