+++ src/main.tsx
@@ -1,17 +1,9 @@
 import * as Sentry from '@sentry/react';
-import { QueryClientProvider } from '@tanstack/react-query';
 import { Analytics } from '@vercel/analytics/react';
-import { lazy, StrictMode, Suspense } from 'react';
+import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import { queryClient } from './lib/queryClient.ts';
-
-// Dev-only: TanStack Query Devtools. Lazy + import.meta.env.DEV guard so the
-// ~40 KB devtools bundle never ships to production.
-const ReactQueryDevtools = import.meta.env.DEV
-  ? lazy(() => import('@tanstack/react-query-devtools').then((m) => ({ default: m.ReactQueryDevtools })))
-  : null;
 
 Sentry.init({
   dsn: import.meta.env.VITE_SENTRY_DSN,
@@ -25,14 +17,7 @@ Sentry.init({
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <QueryClientProvider client={queryClient}>
-      <App />
-      <Analytics />
-      {ReactQueryDevtools && (
-        <Suspense fallback={null}>
-          <ReactQueryDevtools buttonPosition="bottom-left" initialIsOpen={false} />
-        </Suspense>
-      )}
-    </QueryClientProvider>
+    <App />
+    <Analytics />
   </StrictMode>,
 );
